requirements.txt
graia-application-mirai
pyyaml